Clearer port to Core and Core devices
Now we are not using M5Unified for Core2 and Core (basic, Gray and Fire) devices. This change was needed because M5Unified fills the IRAM_tex0 table with functions that limits bruce development.

The libraries M5Core2 and M5Stack were partially copied to the project, and modified to avoid conflicts with TFT_eSPI.

All credits to M5Stack!